Fixed bug where only tensors containing floats could be converted to string.

+++ src/MiniTorchLib/cuda_common.h
@@ -21,7 +21,7 @@ typedef float floatX;
 #elif defined(ENABLE_FP16)
 typedef half floatX;
 #else
-typedef float floatX;
+typedef __nv_bfloat16 floatX;
 #endif
 
 /////////////////////////////
